[#1] Moved and renamed const.js
